refactoring functions and classes out of App.py. Changes to get_time logic
