+++ lib/db.py
@@ -1,0 +1,42 @@
+import sqlite3
+
+DB_PATH = "agent_traces.db"
+
+def init_db():
+    """Initialize SQLite database with required tables"""
+    conn = sqlite3.connect(DB_PATH)
+    cursor = conn.cursor()
+    # Sessions table
+    cursor.execute('''
+        CREATE TABLE IF NOT EXISTS sessions (
+            session_id TEXT PRIMARY KEY,
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+            updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
+        )
+    ''')
+    # Messages table
+    cursor.execute('''
+        CREATE TABLE IF NOT EXISTS messages (
+            message_id TEXT PRIMARY KEY,
+            session_id TEXT NOT NULL,
+            role TEXT NOT NULL CHECK (role IN ('user', 'assistant', 'system')),
+            content TEXT NOT NULL,
+            status TEXT DEFAULT 'completed' CHECK (status IN ('pending', 'running', 'completed', 'error')),
+            created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
+            completed_at TIMESTAMP,
+            FOREIGN KEY (session_id) REFERENCES sessions (session_id)
+        )
+    ''')
+    # Traces table
+    cursor.execute('''
+        CREATE TABLE IF NOT EXISTS traces (
+            id INTEGER PRIMARY KEY AUTOINCREMENT,
+            message_id TEXT NOT NULL,
+            event TEXT NOT NULL,
+            data TEXT NOT NULL,
+            timestamp REAL NOT NULL,
+            FOREIGN KEY (message_id) REFERENCES messages (message_id)
+        )
+    ''')
+    conn.commit()
+    conn.close()
